Added work units
Each task has a number of work units, each of which has a weight. The amount
of work that that task does is proportional to the weight.

The number of work units and the range from which their weights are generated
(using a uniform distribution) can be specified at the command line, e.g.
they're runtime configurable.