Implemented push and pop to/from a work array
This items pushed or popped are also in a work array, so these functions are
less about pushing and popping items as they are about concatenating and
splitting work arrays.

When pushing the work arrays max size may need to be increased, this is done
with realloc. When popping, however, the array size is never reduced.