Changed to "safe" migration
Migrations are now always performed, the sender may send 0 elements if they
are not migrating and the recipient will always accept MAX_MIGRATE elements
from all peers.

In the future we may include "unsafe" migrations and "pre-negociated"
migrations as options as well, but there currently doesn't appear to be a
need.